001-prime-time/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "proto-prime-time"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "BSD 3-Clause License"}
dependencies = [
  "attrs",
  "click"
]
dynamic=["version"]

[project.optional-dependencies]
test = [
  "pytest",
]

[project.scripts]
proto-prime-time = "proto_prime_time.cli:cli"
